+++ terraform/slurm_cluster/variables.tf
@@ -42,92 +42,60 @@ variable "controller_instance_config" {
 Creates a controller instance with given configuration.
 EOD
   type = object({
-    access_config = list(object({
-      nat_ip       = string
-      network_tier = string
-    }))
-    additional_disks = list(object({
-      disk_name    = string
-      device_name  = string
-      disk_size_gb = number
-      disk_type    = string
-      disk_labels  = map(string)
-      auto_delete  = bool
-      boot         = bool
-    }))
-    can_ip_forward         = bool
-    disable_smt            = bool
-    disk_auto_delete       = bool
-    disk_labels            = map(string)
-    disk_size_gb           = number
-    disk_type              = string
-    enable_confidential_vm = bool
-    enable_oslogin         = bool
-    enable_shielded_vm     = bool
-    gpu = object({
+    access_config = optional(list(object({
+      nat_ip       = optional(string)
+      network_tier = optional(string)
+    })), [])
+    additional_disks = optional(list(object({
+      disk_name    = optional(string)
+      device_name  = optional(string)
+      disk_size_gb = optional(number)
+      disk_type    = optional(string)
+      disk_labels  = optional(map(string), {})
+      auto_delete  = optional(bool, true)
+      boot         = optional(bool, false)
+    })), [])
+    can_ip_forward         = optional(bool, false)
+    disable_smt            = optional(bool, false)
+    disk_auto_delete       = optional(bool, true)
+    disk_labels            = optional(map(string), {})
+    disk_size_gb           = optional(number)
+    disk_type              = optional(string, "n1-standard-1")
+    enable_confidential_vm = optional(bool, false)
+    enable_oslogin         = optional(bool, true)
+    enable_shielded_vm     = optional(bool, false)
+    gpu = optional(object({
       count = number
       type  = string
-    })
-    instance_template   = string
-    labels              = map(string)
-    machine_type        = string
-    metadata            = map(string)
-    min_cpu_platform    = string
-    network_ip          = string
-    on_host_maintenance = string
-    preemptible         = bool
-    region              = string
-    service_account = object({
-      email  = string
-      scopes = list(string)
-    })
-    shielded_instance_config = object({
-      enable_integrity_monitoring = bool
-      enable_secure_boot          = bool
-      enable_vtpm                 = bool
-    })
-    source_image_family  = string
-    source_image_project = string
-    source_image         = string
-    static_ip            = string
-    subnetwork_project   = string
-    subnetwork           = string
-    tags                 = list(string)
-    zone                 = string
+    }))
+    instance_template   = optional(string)
+    labels              = optional(map(string), {})
+    machine_type        = optional(string)
+    metadata            = optional(map(string), {})
+    min_cpu_platform    = optional(string)
+    network_ip          = optional(string)
+    on_host_maintenance = optional(string)
+    preemptible         = optional(bool, false)
+    region              = optional(string)
+    service_account = optional(object({
+      email  = optional(string)
+      scopes = optional(list(string), ["https://www.googleapis.com/auth/cloud-platform"])
+    }))
+    shielded_instance_config = optional(object({
+      enable_integrity_monitoring = optional(bool, true)
+      enable_secure_boot          = optional(bool, true)
+      enable_vtpm                 = optional(bool, true)
+    }))
+    source_image_family  = optional(string)
+    source_image_project = optional(string)
+    source_image         = optional(string)
+    static_ip            = optional(string)
+    subnetwork_project   = optional(string)
+    subnetwork           = optional(string)
+    tags                 = optional(list(string), [])
+    zone                 = optional(string)
   })
-  default = {
-    access_config            = []
-    additional_disks         = []
-    can_ip_forward           = false
-    disable_smt              = false
-    disk_auto_delete         = true
-    disk_labels              = {}
-    disk_size_gb             = null
-    disk_type                = null
-    enable_confidential_vm   = false
-    enable_oslogin           = true
-    enable_shielded_vm       = false
-    gpu                      = null
-    instance_template        = null
-    labels                   = {}
-    machine_type             = "n1-standard-1"
-    metadata                 = {}
-    min_cpu_platform         = null
-    network_ip               = null
-    on_host_maintenance      = null
-    preemptible              = false
-    region                   = null
-    service_account          = null
-    shielded_instance_config = null
-    source_image_family      = null
-    source_image_project     = null
-    source_image             = null
-    static_ip                = null
-    subnetwork_project       = null
-    subnetwork               = "default"
-    tags                     = []
-    zone                     = null
-  }
+  default = {}
 }
 
 ######################
@@ -149,37 +117,22 @@ Creates a hybrid controller with given configuration.
 See 'main.tf' for valid keys.
 EOD
   type = object({
-    google_app_cred_path    = string
-    slurm_control_host      = string
-    slurm_control_host_port = string
-    slurm_control_addr      = string
-    slurm_bin_dir           = string
-    slurm_log_dir           = string
-    output_dir              = string
-    install_dir             = string
-    munge_mount = object({
-      server_ip     = string
-      remote_mount  = string
-      fs_type       = string
-      mount_options = string
-    })
+    google_app_cred_path    = optional(string)
+    slurm_control_host      = optional(string)
+    slurm_control_host_port = optional(string)
+    slurm_control_addr      = optional(string)
+    slurm_bin_dir           = optional(string)
+    slurm_log_dir           = optional(string)
+    output_dir              = optional(string)
+    install_dir             = optional(string)
+    munge_mount = optional(object({
+      server_ip     = optional(string)
+      remote_mount  = optional(string, "/etc/munge")
+      fs_type       = optional(string, "nfs")
+      mount_options = optional(string)
+    }), {})
   })
-  default = {
-    google_app_cred_path    = null
-    slurm_control_host      = null
-    slurm_control_host_port = null
-    slurm_control_addr      = null
-    slurm_bin_dir           = "/usr/local/bin"
-    slurm_log_dir           = "/var/log/slurm"
-    output_dir              = "/etc/slurm"
-    install_dir             = null
-    munge_mount = {
-      server_ip     = null
-      remote_mount  = "/etc/munge"
-      fs_type       = "nfs"
-      mount_options = null
-    }
-  }
+  default = {}
 }
 
 #########
@@ -189,60 +142,60 @@ EOD
 variable "login_nodes" {
   description = "List of slurm login instance definitions."
   type = list(object({
-    access_config = list(object({
-      nat_ip       = string
-      network_tier = string
-    }))
-    additional_disks = list(object({
-      disk_name    = string
-      device_name  = string
-      disk_size_gb = number
-      disk_type    = string
-      disk_labels  = map(string)
-      auto_delete  = bool
-      boot         = bool
-    }))
-    can_ip_forward         = bool
-    disable_smt            = bool
-    disk_auto_delete       = bool
-    disk_labels            = map(string)
-    disk_size_gb           = number
-    disk_type              = string
-    enable_confidential_vm = bool
-    enable_oslogin         = bool
-    enable_shielded_vm     = bool
-    gpu = object({
+    access_config = optional(list(object({
+      nat_ip       = optional(string)
+      network_tier = optional(string)
+    })), [])
+    additional_disks = optional(list(object({
+      disk_name    = optional(string)
+      device_name  = optional(string)
+      disk_size_gb = optional(number)
+      disk_type    = optional(string)
+      disk_labels  = optional(map(string), {})
+      auto_delete  = optional(bool, true)
+      boot         = optional(bool, false)
+    })), [])
+    can_ip_forward         = optional(bool, false)
+    disable_smt            = optional(bool, false)
+    disk_auto_delete       = optional(bool, true)
+    disk_labels            = optional(map(string), {})
+    disk_size_gb           = optional(number)
+    disk_type              = optional(string, "n1-standard-1")
+    enable_confidential_vm = optional(bool, false)
+    enable_oslogin         = optional(bool, true)
+    enable_shielded_vm     = optional(bool, false)
+    gpu = optional(object({
       count = number
       type  = string
-    })
+    }))
     group_name          = string
-    instance_template   = string
-    labels              = map(string)
-    machine_type        = string
-    metadata            = map(string)
-    min_cpu_platform    = string
-    network_ips         = list(string)
-    num_instances       = number
-    on_host_maintenance = string
-    preemptible         = bool
-    region              = string
-    service_account = object({
-      email  = string
-      scopes = list(string)
-    })
-    shielded_instance_config = object({
-      enable_integrity_monitoring = bool
-      enable_secure_boot          = bool
-      enable_vtpm                 = bool
-    })
-    source_image_family  = string
-    source_image_project = string
-    source_image         = string
-    static_ips           = list(string)
-    subnetwork_project   = string
-    subnetwork           = string
-    tags                 = list(string)
-    zone                 = string
+    instance_template   = optional(string)
+    labels              = optional(map(string), {})
+    machine_type        = optional(string)
+    metadata            = optional(map(string), {})
+    min_cpu_platform    = optional(string)
+    network_ip          = optional(string)
+    num_instances       = optional(number, 1)
+    on_host_maintenance = optional(string)
+    preemptible         = optional(bool, false)
+    region              = optional(string)
+    service_account = optional(object({
+      email  = optional(string)
+      scopes = optional(list(string), ["https://www.googleapis.com/auth/cloud-platform"])
+    }))
+    shielded_instance_config = optional(object({
+      enable_integrity_monitoring = optional(bool, true)
+      enable_secure_boot          = optional(bool, true)
+      enable_vtpm                 = optional(bool, true)
+    }))
+    source_image_family  = optional(string)
+    source_image_project = optional(string)
+    source_image         = optional(string)
+    static_ips           = optional(list(string), [])
+    subnetwork_project   = optional(string)
+    subnetwork           = optional(string)
+    tags                 = optional(list(string), [])
+    zone                 = optional(string)
   }))
   default = []
 }
@@ -256,142 +209,86 @@ variable "partitions" {
 Cluster partitions as a list. See module slurm_partition.
 EOD
   type = list(object({
-    enable_job_exclusive              = bool
-    enable_placement_groups           = bool
-    partition_conf                    = map(string)
-    partition_startup_scripts_timeout = number
-    partition_startup_scripts = list(object({
+    enable_job_exclusive              = optional(bool, false)
+    enable_placement_groups           = optional(bool, false)
+    partition_conf                    = optional(map(string), {})
+    partition_startup_scripts_timeout = optional(number, 300)
+    partition_startup_scripts = optional(list(object({
       filename = string
       content  = string
-    }))
-    partition_feature = string
+    })), [])
+    partition_feature = optional(string)
     partition_name    = string
-    partition_nodes = list(object({
-      node_count_static      = number
-      node_count_dynamic_max = number
+    partition_nodes = optional(list(object({
+      node_count_static      = optional(number, 0)
+      node_count_dynamic_max = optional(number, 1)
       group_name             = string
-      node_conf              = map(string)
-      additional_disks = list(object({
-        disk_name    = string
-        device_name  = string
-        disk_size_gb = number
-        disk_type    = string
-        disk_labels  = map(string)
-        auto_delete  = bool
-        boot         = bool
-      }))
-      access_config = list(object({
+      node_conf              = optional(map(string), {})
+      additional_disks = optional(list(object({
+        disk_name    = optional(string)
+        device_name  = optional(string)
+        disk_size_gb = optional(number)
+        disk_type    = optional(string)
+        disk_labels  = optional(map(string), {})
+        auto_delete  = optional(bool, true)
+        boot         = optional(bool, false)
+      })), [])
+      access_config = optional(list(object({
         network_tier = string
-      }))
-      bandwidth_tier         = string
-      can_ip_forward         = bool
-      disable_smt            = bool
-      disk_auto_delete       = bool
-      disk_labels            = map(string)
-      disk_size_gb           = number
-      disk_type              = string
-      enable_confidential_vm = bool
-      enable_oslogin         = bool
-      enable_shielded_vm     = bool
-      enable_spot_vm         = bool
-      gpu = object({
+      })), [])
+      bandwidth_tier         = optional(string, "platform_default")
+      can_ip_forward         = optional(bool, false)
+      disable_smt            = optional(bool, false)
+      disk_auto_delete       = optional(bool, true)
+      disk_labels            = optional(map(string), {})
+      disk_size_gb           = optional(number)
+      disk_type              = optional(string)
+      enable_confidential_vm = optional(bool, false)
+      enable_oslogin         = optional(bool, true)
+      enable_shielded_vm     = optional(bool, false)
+      enable_spot_vm         = optional(bool, false)
+      gpu = optional(object({
         count = number
         type  = string
-      })
-      instance_template   = string
-      labels              = map(string)
-      machine_type        = string
-      metadata            = map(string)
-      min_cpu_platform    = string
-      on_host_maintenance = string
-      preemptible         = bool
-      service_account = object({
-        email  = string
-        scopes = list(string)
-      })
-      shielded_instance_config = object({
-        enable_integrity_monitoring = bool
-        enable_secure_boot          = bool
-        enable_vtpm                 = bool
-      })
-      spot_instance_config = object({
-        termination_action = string
-      })
-      source_image_family  = string
-      source_image_project = string
-      source_image         = string
-      tags                 = list(string)
-    }))
-    network_storage = list(object({
+      }))
+      instance_template   = optional(string)
+      labels              = optional(map(string), {})
+      machine_type        = optional(string)
+      metadata            = optional(map(string), {})
+      min_cpu_platform    = optional(string)
+      on_host_maintenance = optional(string)
+      preemptible         = optional(bool, false)
+      service_account = optional(object({
+        email  = optional(string)
+        scopes = optional(list(string), ["https://www.googleapis.com/auth/cloud-platform"])
+      }))
+      shielded_instance_config = optional(object({
+        enable_integrity_monitoring = optional(bool, true)
+        enable_secure_boot          = optional(bool, true)
+        enable_vtpm                 = optional(bool, true)
+      }))
+      spot_instance_config = optional(object({
+        termination_action = optional(string, "STOP")
+      }))
+      source_image_family  = optional(string)
+      source_image_project = optional(string)
+      source_image         = optional(string)
+      tags                 = optional(list(string), [])
+    })), [])
+    network_storage = optional(list(object({
       local_mount   = string
       fs_type       = string
       server_ip     = string
       remote_mount  = string
       mount_options = string
-    }))
-    region             = string
-    subnetwork_project = string
-    subnetwork         = string
-    zone_target_shape  = string
-    zone_policy_allow  = list(string)
-    zone_policy_deny   = list(string)
+    })), [])
+    region             = optional(string)
+    subnetwork_project = optional(string)
+    subnetwork         = optional(string)
+    zone_target_shape  = optional(string, "ANY_SINGLE_ZONE")
+    zone_policy_allow  = optional(list(string), [])
+    zone_policy_deny   = optional(list(string), [])
   }))
-  default = [
-    {
-      enable_job_exclusive    = false
-      enable_placement_groups = false
-      partition_conf = {
-        Default = "YES"
-      }
-      partition_startup_scripts_timeout = 300
-      partition_startup_scripts         = []
-      partition_feature                 = null
-      partition_name                    = "debug"
-      partition_nodes = [
-        {
-          node_count_static        = 20
-          node_count_dynamic_max   = 0
-          group_name               = "test"
-          node_conf                = {}
-          additional_disks         = []
-          access_config            = []
-          bandwidth_tier           = "platform_default"
-          can_ip_forward           = false
-          disable_smt              = false
-          disk_auto_delete         = true
-          disk_labels              = {}
-          disk_size_gb             = null
-          disk_type                = null
-          enable_confidential_vm   = false
-          enable_oslogin           = true
-          enable_shielded_vm       = false
-          enable_spot_vm           = false
-          gpu                      = null
-          instance_template        = null
-          labels                   = {}
-          machine_type             = "n1-standard-1"
-          metadata                 = {}
-          min_cpu_platform         = null
-          on_host_maintenance      = null
-          preemptible              = false
-          service_account          = null
-          shielded_instance_config = null
-          spot_instance_config     = null
-          source_image_family      = null
-          source_image_project     = null
-          source_image             = null
-          tags                     = []
-        }
-      ]
-      network_storage    = []
-      region             = null
-      subnetwork_project = null
-      subnetwork         = "default"
-      zone_target_shape  = "ANY_SINGLE_ZONE"
-      zone_policy_allow  = []
-      zone_policy_deny   = []
-    },
-  ]
 
   validation {
     condition     = length(var.partitions) > 0
@@ -471,19 +368,13 @@ EOD
 variable "cloud_parameters" {
   description = "cloud.conf options."
   type = object({
-    no_comma_params = bool
-    resume_rate     = number
-    resume_timeout  = number
-    suspend_rate    = number
-    suspend_timeout = number
+    no_comma_params = optional(bool, false)
+    resume_rate     = optional(number, 0)
+    resume_timeout  = optional(number, 300)
+    suspend_rate    = optional(number, 0)
+    suspend_timeout = optional(number, 300)
   })
-  default = {
-    no_comma_params = false
-    resume_rate     = 0
-    resume_timeout  = 300
-    suspend_rate    = 0
-    suspend_timeout = 300
-  }
+  default = {}
 }
 
 variable "disable_default_mounts" {