+++ terraform/slurm_cluster/versions.tf
@@ -15,7 +15,7 @@
  */
 
 terraform {
-  required_version = "~> 1.0"
+  required_version = "~> 1.3"
 
   required_providers {
     random = "~> 3.0"